Refactor completers to separate files
