+++ README.md
@@ -131,8 +131,10 @@ openai.api_key = '<API Key>'
 import pandas_gpt
 pandas_gpt.completer = pandas_gpt.OpenAI(
   model='gpt-3.5-turbo',
-  engine='<Engine>',
-  deployment_id='<Deployment ID>',
+  client_config = dict(
+    engine='<Engine>',
+    deployment_id='<Deployment ID>',
+  ),
 )
 ```
