+++ pandas_gpt/completers/__init__.py
@@ -1,0 +1,5 @@
+from pandas_gpt.completers.litellm import LiteLLM
+from pandas_gpt.completers.openai import OpenAI
+from pandas_gpt.completers.openrouter import OpenRouter
+
+__all__ = ["LiteLLM", "OpenAI", "OpenRouter"]
