+++ pandas_gpt/completers/openai.py
@@ -1,6 +1,10 @@
 from dataclasses import dataclass
+import os
+from typing import Any
 
-__all__ = ["OpenAI", "LiteLLM", "OpenRouter"]
+__all__ = ["OpenAI"]
+
+default_system_prompt = "Write the function in a Python code block with all necessary imports and no example usage."
 
 
 @dataclass
@@ -10,9 +14,18 @@ class OpenAI:
     _cache: dict[str, str]
     _client: Any
 
-    def __init__(self, model: str, **client_config):
+    def __init__(
+        self,
+        model: str,
+        *,
+        system_prompt: str | None = default_system_prompt,
+        client_config: dict[str, Any] | None = None,
+    ):
         self.completion_config = {"model": model}
-        self.client_config = client_config
+        self.client_config = client_config if client_config is not None else {}
+        self.system_prompt = (
+            system_prompt if system_prompt is not None else default_system_prompt
+        )
         self._cache = {}
         self._client = None
 
@@ -20,7 +33,7 @@ class OpenAI:
         completion = self._cache.get(prompt) or self.run_completion_function(
             **self.completion_config,
             messages=[
-                dict(role="system", content=system_prompt),
+                dict(role="system", content=self.system_prompt),
                 dict(role="user", content=prompt),
             ],
         )
@@ -41,20 +54,3 @@ class OpenAI:
                 client_config["api_key"] = api_key
             self._client = openai.OpenAI(**client_config)
         return self._client.chat.completions.create(**kw)
-
-
-@dataclass
-class OpenRouter(OpenAI):
-    def __init__(self, model: str, **config):
-        super().__init__(self, model, base_url="https://openrouter.ai/api/v1", **config)
-
-
-class LiteLLM(OpenAI):
-    def run_completion_function(self, **kw):
-        try:
-            import litellm
-        except ImportError:
-            raise Exception(
-                "The package `litellm` could not be found. You can fix this error by running `pip install pandas-gpt[litellm]` or passing a custom `completer` argument."
-            )
-        return litellm.completion(**kw)
